fix: medicine order int to float

+++ src/models/models.py
@@ -132,11 +132,11 @@ class MedicineOrder(BaseModel):
     form = Column(String(100), nullable=True)
     strength = Column(String(100), nullable=True)
     pharmaceuticals = Column(String(100), nullable=True)
-    quantity = Column(Integer, nullable=False)
-    unit_price_tp = Column(Integer, nullable=True)
-    unit_price_mrp = Column(Integer, nullable=False)
-    unit_discount_percent = Column(Integer, nullable=False)
-    total = Column(Integer, nullable=False)
+    quantity = Column(Float, nullable=False)
+    unit_price_tp = Column(Float, nullable=True)
+    unit_price_mrp = Column(Float, nullable=False)
+    unit_discount_percent = Column(Float, nullable=False)
+    total = Column(Float, nullable=False)
 
 
 ################
